+++ tests/test_documentation.py
@@ -21,9 +21,10 @@ OTHER DEALINGS IN THE SOFTWARE.
 """
 import json
 
-import hug
 from falcon import Request
 from falcon.testing import StartResponseMock, create_environ
+
+import hug
 
 api = hug.API(__name__)